requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
pyyaml>=6.0
prettytable>=3.17.0
